@@ -1,34 +1,28 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import './UserProfile.module.css';
-import axios from 'axios';
-import Cookies from 'js-cookie';
-import { useEffect } from 'react';
+import './userProfile.css';
+import { useState } from 'react';
 
-const UserProfile = (props) => {
-  const isAuthenticated = useSelector(state => state.auth.isAuthenticated);
+
+const UserProfile = () => {
   const counter = useSelector(state => state.counter.counter);
 
-  useEffect(async () => {
-    const headers = {
-      'Authorization': `Bearer ${Cookies.get('user')}`
-    }
-    try {
-      const response = await axios.get('/api/v1/products', headers);
-      console.log(response);
-    } catch (err) {
-      console.log(err)
-    }
-  }, []);
+  const [productsState, setProductsState] = useState(
+    [
+      { id: 1, name: "iPhone 13", price: 3000 },
+      { id: 2, name: "iPhone 12", price: 3000 },
+      { id: 3, name: "galaxy s20", price: 3000 }
+    ]
+  );
+
 
   return (
     <React.Fragment>
       <main className="profile">
-        {/* {isAuthenticated ? null : props.history.push("/login")} */}
         <h2>My User Profile</h2>
         <p>your current count is: {counter}</p>
+        {/* {<Products products={productsState}></Products>} */}
       </main>
-
     </React.Fragment>
 
   );
